Bug fixes1

+++ models/tasks.js
@@ -1,15 +1,15 @@
 const db = require('../database/db');
 
 const Tasks = {
-    getAll: () => {
-        const query = 'SELECT * FROM tasks';
-        return db.query(query).then((response) => {
+    getAll: (user_id) => {
+        const query = 'SELECT * FROM tasks WHERE user_id = $1';
+        return db.query(query, [user_id]).then((response) => {
             return response.rows;
         });
     },
-    getById: (id) => {
-        const query = `SELECT * FROM tasks WHERE id = ${id}`;
-        return db.query(query).then((response) => {
+    getById: (id, user_id) => {
+        const query = `SELECT * FROM tasks WHERE id = $1 and user_id = $2`;
+        return db.query(query, [id, user_id]).then((response) => {
             return response.rows ? response.rows[0] : {};
         });
     },
@@ -21,12 +21,12 @@ const Tasks = {
                 return response.rows ? response.rows[0] : {};
             });
     },
-    delete: (id) => {
-        const query = `DELETE FROM tasks WHERE id = $1`;
-        return db.query(query, [id]);
+    delete: (id, user_id) => {
+        const query = `DELETE FROM tasks WHERE id = $1 and user_id = $2`;
+        return db.query(query, [id, user_id]);
     },
     update: ({ id, name, description, priority, due_date, user_id }) => {
-        const query = `UPDATE tasks SET name = $2, description = $3, priority = $4, due_date = $5, user_id = $6 WHERE id = $1 RETURNING *`;
+        const query = `UPDATE tasks SET name = $2, description = $3, priority = $4, due_date = $5  WHERE id = $1 and user_id = $6 RETURNING *`;
         return db
             .query(query, [id, name, description, priority, due_date, user_id])
             .then((response) => {
